+++ include/class/Sync.php
@@ -38,6 +38,14 @@ class Sync extends Daemon
 			return;
 		}
 
+        $height = Block::getHeight();
+        if($height >= DELETE_CHAIN_HEIGHT) {
+            $diff = DELETE_CHAIN_HEIGHT - $height;
+            _log("Pop $diff blocks at demand");
+            Block::pop($diff);
+            return;
+        }
+
 		Peer::deleteDeadPeers();
 		Peer::blacklistInactivePeers();
 		Peer::blacklistIncompletePeers();
